[#S07P31A102-223, #S07P31A102-224] fix : UPDATE, DELETE meetup 관리자체크

+++ backend/src/main/java/com/meetup/backend/service/meetup/MeetupServiceImpl.java
@@ -60,16 +60,17 @@ public class MeetupServiceImpl implements MeetupService {
     public void updateMeetup(MeetupUpdateRequestDto meetupUpdateRequestDto, String userId, Long meetupId) {
 
         Meetup meetup = meetupRepository.findById(meetupId).orElseThrow(() -> new ApiException(ExceptionEnum.MEETUP_NOT_FOUND));
-
-        meetup.changeMeetup(meetupUpdateRequestDto.getTitle(), meetupUpdateRequestDto.getColor());
+        if (meetup.getManager().getId().equals(userId))
+            meetup.changeMeetup(meetupUpdateRequestDto.getTitle(), meetupUpdateRequestDto.getColor());
 
     }
 
     @Override
     @Transactional
-    public void deleteMeetup(Long meetupId) {
+    public void deleteMeetup(Long meetupId, String userId) {
         Meetup meetup = meetupRepository.findById(meetupId).orElseThrow(() -> new ApiException(ExceptionEnum.MEETUP_NOT_FOUND));
-        meetup.deleteMeetup(true);
+        if (meetup.getManager().getId().equals(userId))
+            meetup.deleteMeetup(true);
     }
 
     @Override
